apply css

+++ src/controllers/videoController.js
@@ -1,14 +1,16 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 //Video.find({}, (error, video) =>{});
 export const home = async(req, res) => {
-    const videos = await Video.find({}).sort({createdAt : "asc"});
+    const videos = await Video.find({}).sort({createdAt : "asc"}).populate("owner");
     return res.render("home", {pageTitle:"Home", videos }); 
 }
 
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
+
     if(!video){
         return res.status(404).render("404", {pageTitle : "Video not found"});
     }
@@ -16,8 +18,14 @@ export const watch = async (req, res) => {
     
 }
 export const getEdit = async (req, res) => {
+    const {user : {_id}} = req.session
     const { id } = req.params;
     const video = await Video.findById(id);
+
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
+
     if(!video){
         return res.status(404).render("404", {pageTitle : "Video not found"});  
     }
@@ -25,9 +33,14 @@ export const getEdit = async (req, res) => {
 }
 
 export const postEdit = async (req, res) => {
+    const {user : {_id}} = req.session
     const { id } = req.params;
     const {title, description, hashtags} = req.body;
     const video = await Video.exists({_id : id});
+
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     
     if(!video){
         return res.status(404).render("404", {pageTitle : "Video not found"});
@@ -45,15 +58,22 @@ export const getUpload = (req, res) => {
 }
 
 export const postUpload = async (req, res) => {
+    const {user : {_id}} = req.session;
     const {path : fileUrl} = req.file;
     const {title, description, hashtags} = req.body; 
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title : title,
+            owner : _id, 
             fileUrl,
             description : description,
             hashtags : Video.formatHashtags(hashtags),
         });
+
+        const user = await User.findById(_id);
+        user.videos.push(newVideo);
+        user.save();
+
         return res.redirect("/");
     }
     catch(error){
@@ -63,8 +83,17 @@ export const postUpload = async (req, res) => {
 }
 
 export const deleteVideo = async (req, res) => {
+    const {user : {_id}} = req.session;
     const {id} = req.params;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render("404", {pageTitle : "Video not found"});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status("403").redirect("/");
+    }
     await Video.findByIdAndDelete(id);
+    
     return res.redirect("/");
 }
 
@@ -76,7 +105,7 @@ export const search = async (req, res) => {
             title : {
                 $regex: new RegExp(keyword, "i")
             },
-        });
+        }).populate("owner");
     }
     return res.render("search", {pageTitle : "Search", videos});
 }
